clean in progress

+++ start.sh
@@ -14,19 +14,29 @@ printf "\033[1;31mclean\n\033[0m"
 NB_RUN=$(docker ps -aq | wc -l)
 NB_IMG=$(docker images -aq | wc -l)
 
-if [[ $NB_RUN != "0" ]]
+
+if [[ $1 == "clean" ]]
 then
-	docker stop $(docker ps -aq)
-	docker rm $(docker ps -aq)
-	docker system prune
-else
-	printf "no containers running\n"
+	if [[ $NB_RUN != "0" ]]
+	then
+		docker stop $(docker ps -aq)
+		docker rm $(docker ps -aq)
+		docker system prune
+	else
+		printf "no containers running\n"
+	fi
+	if [[ $NB_IMG != "0" ]]
+	then
+		docker rmi $(docker images -aq)
+	else
+		printf "no images\n"
+	fi
 fi
-if [[ $NB_IMG != "0" ]]
+
+if [[ $1 == "run" ]]
 then
-	docker rmi $(docker images -aq)
-else
-	printf "no images\n"
+	printf "\033[1;33mrunning\n\033[0m"
+	docker run -it --init eval
 fi
 
 if [[ $1 == "build" ]]
@@ -34,5 +44,5 @@ then
 	printf "\033[1;33mbuild\n\033[0m"
 	docker build -t eval .
 	printf "\033[1;32mexec\n\033[0m"
-	docker run -it --init --rm --name="eval" eval
+	docker run -it --init eval
 fi
